requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.30.6
Flask==3.0.0
Flask-CORS==4.0.0
SQLAlchemy==2.0.36
psycopg[binary,pool]==3.2.3
psycopg2-binary
alembic==1.13.2
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.9
PyJWT==2.9.0
passlib[bcrypt]==1.7.4
python-socketio==5.11.4
httpx==0.27.2
aiohttp==3.10.11
tenacity==9.0.0
geoalchemy2==0.14.7
shapely==2.0.5
reportlab==4.2.2
opentelemetry-api==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation-fastapi==0.46b0
PyYAML==6.0.2
